test: update test codes for new logic
